Added background grid if there is no movie. Made it so that no movie is required in ProjectSetup. Updated plotting and converting for Louis lab data.

+++ perframe/ConvertLarvaeLouis2JAABA.m
@@ -23,34 +23,32 @@ trxunits = {
 
 %% parse parameters
 
-[inmoviefile,indatafile,inconfigfile,inkinmatfile,expi,...
+[inmoviefile,indatafile,inkinmatfile,expi,...
   expdir,moviefilestr,trxfilestr,perframedirstr,...
   arenatype,arenacenterx,arenacentery,...
   arenaradius,arenawidth,arenaheight,...
   pxpermm,fps,overridefps,overridearena,...
-  dosoftlink] = myparse(varargin,...
-  'inmoviefile','','indatafile','','inconfigfile','','inkinmatfile','','expi',[],...
+  dosoftlink,ncontourpts] = myparse(varargin,...
+  'inmoviefile','','indatafile','','inkinmatfile','','expi',[],...
   'expdir','','moviefilestr','movie.avi','trxfilestr','trx.mat','perframedirstr','perframe',...
   'arenatype','None','arenacenterx',0,'arenacentery',0,...
   'arenaradius',123,'arenawidth',123,'arenaheight',123,...
   'pxpermm',1,'fps',30,...
   'overridefps',false,'overridearena',false,...
-  'dosoftlink',false); %#ok<ASGLU>
+  'dosoftlink',false,...
+  'ncontourpts',20); %#ok<ASGLU>
 
 % get experiment_name from full path
 [~,experiment_name] = myfileparts(expdir); %#ok<NASGU>
 
 % check that required inputs exist
-if ~exist(inmoviefile,'file'),
+ismovie = ~isempty(inmoviefile);
+if ismovie && ~exist(inmoviefile,'file'),
   msg = sprintf('Input movie file %s does not exist',inmoviefile);
   return;
 end
 if ~exist(indatafile,'file'),
   msg = sprintf('Input data file %s does not exist',indatafile);
-  return;
-end
-if ~exist(inconfigfile,'file'),
-  msg = sprintf('Input config file %s does not exist',inconfigfile);
   return;
 end
 if ~exist(inkinmatfile','file'),
@@ -77,74 +75,58 @@ while true,
   if ~ischar(s),
     break;
   end
-  ss = strsplit(s,',');
-  ss = str2num(char(ss(1:17))); %#ok<ST2NM>
+  s = strtrim(s);
+  if isempty(s),
+    continue;
+  end
+  ss = str2double(strsplit(s,','));
   indata.data(:,end+1) = ss;
 end
 %indata.data = [(0:size(indata.data,1)-1)',indata.data];
 fclose(fid);
 
-% read in relevant parameters from config file
-fid = fopen(inconfigfile,'r');
-readpxpermm = overridearena;
-readtickpermm = false;
-while true,
-  s = fgetl(fid);
-  if ~ischar(s),
-    break;
-  end
-  s = strtrim(s);
-  if isempty(s),
-    continue;
-  end
-  if ~readpxpermm,
-    m = regexp(s,'^Camera Calibration \(um per pixel\): (.*)$','tokens','once');
-    if ~isempty(m),
-      pxpermm = 1000/str2double(m{1});
-      readpxpermm = true;
-    end
-  end
-  if ~readtickpermm,
-    m = regexp(s,'^Stage Calibration \(tick per mm\): x = (.*), y= (.*)$','tokens','once');
-    if ~isempty(m),
-      xtickpermm = str2double(m{1});
-      ytickpermm = str2double(m{2});
-      readtickpermm = true;
-    end
-  end
-end
-fclose(fid);
-if ~readtickpermm,
-  msg = sprintf('Could not read stage ticks per mm from %s',inconfigfile);
-  return;
-end
-msg{end+1} = sprintf('pxpermm = %f, x-ticks per mm = %f, y-ticks per mm = %f',pxpermm,xtickpermm,ytickpermm);
+% % read in relevant parameters from config file
+% fid = fopen(inconfigfile,'r');
+% if ~ismovie,
+%   pxpermm = 1;
+% end
+% readpxpermm = overridearena || ~ismovie;
+% readtickpermm = false;
+% while true,
+%   s = fgetl(fid);
+%   if ~ischar(s),
+%     break;
+%   end
+%   s = strtrim(s);
+%   if isempty(s),
+%     continue;
+%   end
+%   if ~readpxpermm,
+%     m = regexp(s,'^Camera Calibration \(um per pixel\): (.*)$','tokens','once');
+%     if ~isempty(m),
+%       pxpermm = 1000/str2double(m{1});
+%       readpxpermm = true;
+%     end
+%   end
+%   if ~readtickpermm,
+%     m = regexp(s,'^Stage Calibration \(tick per mm\): x = (.*), y= (.*)$','tokens','once');
+%     if ~isempty(m),
+%       xtickpermm = str2double(m{1});
+%       ytickpermm = str2double(m{2});
+%       readtickpermm = true;
+%     end
+%   end
+% end
+% fclose(fid);
+% if ~readtickpermm,
+%   msg = sprintf('Could not read stage ticks per mm from %s',inconfigfile);
+%   return;
+% end
+% msg{end+1} = sprintf('pxpermm = %f, x-ticks per mm = %f, y-ticks per mm = %f',pxpermm,xtickpermm,ytickpermm);
 
 timestamps = indata.data(2,:);
 
-trx = struct;
-trx.x = indata.data(10,:);
-trx.y = indata.data(11,:);
-trx.xspine = indata.data([3,5,7],:);
-trx.yspine = indata.data([4,6,8],:);
-trx.a = indata.data(9,:) / 4;
-trx.b = trx.a;
-trx.theta = zeros(size(trx.a));
-trx.dt = diff(timestamps);
-trx.id = 1; % only one larva
-trx.moviename = moviefile;
-trx.firstframe = 1;
-trx.endframe = numel(trx.x);
-trx.nframes = numel(trx.x);
-trx.off = 0;
-trx.timestamps = timestamps;
-trx.dt = diff(timestamps);
-
-stagex = indata.data(14,:) / xtickpermm;
-stagey = indata.data(15,:) / ytickpermm;
-
-trx.stagex = stagex * pxpermm;
-trx.stagey = stagey * pxpermm;
+nframes = size(indata.data,2);
 
 % load in kinData
 try
@@ -157,12 +139,12 @@ end
 canmatch = false(1,numel(kinData)); %#ok<NODEF>
 for i = 1:numel(kinData),
   if isfield(kinData{i},'centroidposition'),
-    canmatch(i) = trx.nframes == size(kinData{i}.centroidposition,1);
+    canmatch(i) = nframes == size(kinData{i}.centroidposition,1);
   end
 end
 
 if ~any(canmatch),
-  msg = sprintf('No cells in kinData have the right number of frames (%d)',trx.nframes);
+  msg = sprintf('No cells in kinData have the right number of frames (%d)',nframes);
   return;
 end  
 
@@ -185,12 +167,75 @@ end
 
 kinData = kinData{expi};
 
+trx = struct;
+
+if ismovie,
+  trx.x = indata.data(10,:);
+  trx.y = indata.data(11,:);
+  trx.xspine = indata.data([3,5,7],:);
+  trx.yspine = indata.data([4,6,8],:);
+  trx.a = indata.data(9,:) / 4;
+  trx.b = trx.a;
+  trx.theta = indata.data(13,:)*pi/180;
+else
+  trx.x = kinData.centroidposition(:,1)';
+  trx.y = kinData.centroidposition(:,2)';
+  trx.xspine = [kinData.headposition(:,1)';kinData.centroidposition(:,1)';kinData.tailposition(:,1)'];
+  trx.yspine = [kinData.headposition(:,2)';kinData.centroidposition(:,2)';kinData.tailposition(:,2)'];
+  trx.a = kinData.skeL' / 4;
+  trx.b = trx.a;
+  trx.theta = kinData.bodyangle'*pi/180;
+end
+trx.dt = diff(timestamps);
+trx.id = 1; % only one larva
+trx.moviename = moviefile;
+trx.firstframe = 1;
+trx.endframe = numel(trx.x);
+trx.nframes = numel(trx.x);
+trx.off = 0;
+trx.timestamps = timestamps;
+trx.dt = diff(timestamps);
+
+stagex = indata.data(14,:);
+stagey = indata.data(15,:);
+
+% regress to find scaling from pixels to mm
+[xcoeffs,~,resid] = regress(trx.x(:),[indata.data(14,:)',indata.data(10,:)',ones(nframes,1)]);
+maxerr = max(abs(resid));
+if maxerr > .01,
+  msg = sprintf('Could not find a low error regression from pixels to mms for x, maximum residual = %f with x_mm = %f*stagex + %f*x_px + %f',maxerr,xcoeffs(1),xcoeffs(2),xcoeffs(3));
+  return;
+end
+msg{end+1} = sprintf('x-coordinate px to mm regression has error %f: x_mm = %f*stagex + %f*x_px + %f',maxerr,xcoeffs(1),xcoeffs(2),xcoeffs(3));
+
+[ycoeffs,~,resid] = regress(trx.y(:),[indata.data(15,:)',indata.data(11,:)',ones(nframes,1)]);
+maxerr = max(abs(resid));
+if maxerr > .01,
+  msg = sprintf('Could not find a low error regression from pixels to mms for y, maximum residual = %f with y_mm = %f*stagey + %f*y_px + %f',maxerr,ycoeffs(1),ycoeffs(2),ycoeffs(3));
+  return;
+end
+msg{end+1} = sprintf('y-coordinate px to mm regression has error %f: y_mm = %f*stagey + %f*y_px + %f',maxerr,ycoeffs(1),ycoeffs(2),ycoeffs(3));
+
+pxpermm = (xcoeffs(2)+ycoeffs(2))/2;
+
+if ismovie,
+  contours_px = getContoursLarvaeLouis(indata.data(19:end,:)',ncontourpts);
+else
+  contours_px0 = getContoursLarvaeLouis(indata.data(19:end,:)',ncontourpts);
+  contours_px = nan(size(contours_px0));
+  contours_px(:,:,1) = bsxfun(@plus,indata.data(14,:)'*xcoeffs(1),contours_px0(:,:,1)*xcoeffs(2)) + xcoeffs(3);
+  contours_px(:,:,2) = bsxfun(@plus,indata.data(15,:)'*ycoeffs(1),contours_px0(:,:,2)*ycoeffs(2)) + ycoeffs(3);
+  
+end
 
 trx.x_mm = kinData.centroidposition(:,1)';
 trx.y_mm = kinData.centroidposition(:,2)';
-trx.a_mm = trx.a / pxpermm;
-trx.b_mm = trx.b / pxpermm;
-trx.theta_mm = trx.theta / pxpermm;
+trx.a_mm = kinData.skeL';
+trx.b_mm = kinData.skeL';
+trx.theta_mm = trx.theta;
+
+trx.xcontour = mat2cell(contours_px(:,:,1)',ncontourpts,ones(1,trx.nframes));
+trx.ycontour = mat2cell(contours_px(:,:,2)',ncontourpts,ones(1,trx.nframes));
 
 %% convert to mm
 
@@ -224,8 +269,9 @@ trx = SetLandmarkParameters(trx,arenatype,arenacenterx_mm,arenacentery_mm,...
 
 % create the experiment directory
 if ~exist(expdir,'dir'),
-  [success1,msg] = mkdir(expdir);
+  [success1,msg1] = mkdir(expdir);
   if ~success1,
+    msg = msg1;
     return;
   end
 end
@@ -233,8 +279,9 @@ end
 %% create per-frame directory
 
 if ~exist(perframedir,'dir'),
-  [success1,msg] = mkdir(perframedir);
+  [success1,msg1] = mkdir(perframedir);
   if ~success1,
+    msg = msg1;
     return;
   end
 end
@@ -301,8 +348,8 @@ for i = 1:numel(perframefns2),
 end
 
 data = {stagex}; %#ok<NASGU>
-units = parseunits('mm'); %#ok<NASGU>
-outfile = fullfile(perframedir,'stagex_mm.mat');
+units = parseunits('unit'); %#ok<NASGU>
+outfile = fullfile(perframedir,'stagex.mat');
 try
   save(outfile,'data','units');
 catch ME,
@@ -310,7 +357,7 @@ catch ME,
   return;
 end
 data = {stagey}; %#ok<NASGU>
-outfile = fullfile(perframedir,'stagey_mm.mat');
+outfile = fullfile(perframedir,'stagey.mat');
 try
   save(outfile,'data','units');
 catch ME,
@@ -350,7 +397,7 @@ end
 
 %% copy/soft-link movie
 
-if ~isempty(inmoviefile),
+if ismovie,
   
   if strcmp(fullfile(inmoviefile),fullfile(moviefile)),
     fprintf('Input and out movie files are the same, not copying/linking.\n');
@@ -358,14 +405,37 @@ if ~isempty(inmoviefile),
     
     if dosoftlink,
       if isunix,
-        cmd = sprintf('ln -s %s %s',inmoviefile,moviefile);
-        unix(cmd);
-        % test to make sure it worked
-        [status,result] = unix(sprintf('readlink %s',moviefile));
-        result = strtrim(result);
-        if status ~= 0 || ~strcmp(result,inmoviefile),
-          warndlg(sprintf('Failed to make soft link, copying %s to %s instead',inmoviefile,moviefile));
-          dosoftlink = false;
+        
+        % delete existing files
+        if exist(inmoviefile,'file') && strcmp(inmoviefile,moviefile),
+        elseif exist(inmoviefile,'file'),
+          [status,result] = unix(sprintf('readlink %s',moviefile));
+          result = strtrim(result);
+          if status == 0,
+            msg{end+1} = sprintf('Softlink to %s existed, deleted',result);
+            unix(sprintf('rm %s',moviefile));
+          else
+            res = questdlg(sprintf('File %s exists. Delete and overwrite with softlink to %s?',moviefile,inmoviefile),'Overwrite?','Yes','No','Cancel','Cancel');
+            if strcmpi(res,'No'),
+              msg = sprintf('File %s already existed',moviefile);
+              return;
+            elseif strcmpi(res,'Yes'),
+              msg{end+1} = sprintf('File %s existed and was deleted',moviefile);
+              delete(moviefile);
+            end
+          end
+        end
+
+        if ~exist(moviefile,'file'),
+          cmd = sprintf('ln -s %s %s',inmoviefile,moviefile);
+          unix(cmd);
+          % test to make sure it worked
+          [status,result] = unix(sprintf('readlink %s',moviefile));
+          result = strtrim(result);
+          if status ~= 0 || ~strcmp(result,inmoviefile),
+            warndlg(sprintf('Failed to make soft link, copying %s to %s instead',inmoviefile,moviefile));
+            dosoftlink = false;
+          end
         end
       elseif ispc,
         cmd = sprintf('mkshortcut.vbs /target:"%s" /shortcut:"%s"',inmoviefile,moviefile);
